task_planner/agent.py
"""
GenAI Task Planner — Multi-Agent Sequential Pipeline

A 4-agent pipeline using Google ADK:
  Researcher → Planner → Writer → Reviewer

Each agent passes its output to the next via shared session state (output_key).
"""

from google.adk.agents import LlmAgent, SequentialAgent
from .tools import search_web, get_current_time, save_to_file

# ─── Model Configuration ─────────────────────────────────────────────
MODEL_ID = "gemini-2.0-flash"

# ─── Agent 1: Researcher ─────────────────────────────────────────────
researcher_agent = LlmAgent(
    name="researcher",
    model=MODEL_ID,
    instruction="""You are an expert **Research Analyst**. Your job is to gather comprehensive 
information about the user's task or topic.

**Your Process:**
1. Analyze the user's request to understand what they need
2. Use the `search_web` tool to gather relevant information
3. Synthesize your findings into a well-structured research brief

**Your Output Must Include:**
- **Topic Overview**: What this is about and why it matters
- **Key Findings**: Important facts, data points, and insights (at least 5-7 findings)
- **Current Trends**: What's happening in this space right now
- **Best Practices**: Industry standards and recommended approaches
- **Challenges & Risks**: Potential obstacles and considerations
- **Opportunities**: Areas of potential and growth

**Format**: Use clear markdown with headers, bullet points, and bold key terms.
Be thorough, factual, and specific — your research is the foundation for the entire plan.""",
    description="Researches and gathers comprehensive information about the user's task",
    tools=[search_web],
    output_key="research_output",
)

# ─── Agent 2: Planner ────────────────────────────────────────────────
planner_agent = LlmAgent(
    name="planner",
    model=MODEL_ID,
    instruction="""You are an expert **Strategic Planner**. Based on the research provided, 
create a detailed, actionable task plan.

**Research to build upon:**
{research_output}

**Your Process:**
1. Use the `get_current_time` tool to get today's date for timeline planning
2. Analyze the research findings thoroughly
3. Create a structured, prioritized action plan

**Your Output Must Include:**

## Executive Summary
A 2-3 sentence overview of the plan.

## Goals & Objectives
- Primary goal
- 3-5 specific, measurable objectives

## Task Breakdown
For EACH major task, provide:
| # | Task | Priority | Estimated Effort | Dependencies |
|---|------|----------|-----------------|--------------|
| 1 | ...  | High/Med/Low | X hours/days | None/Task # |

## Detailed Steps
For each task, provide:
- **What**: Clear description of the work
- **How**: Specific approach and methodology
- **Deliverable**: What the output of this step looks like

## Timeline
A realistic phased timeline (Phase 1, Phase 2, etc.)

## Risk Mitigation
Top 3 risks and how to handle them

## Resource Requirements
What tools, skills, or resources are needed

**Be specific, realistic, and actionable.** Every step should be clear enough that someone could start working on it immediately.""",
    description="Creates a detailed, actionable task plan from the research",
    tools=[get_current_time],
    output_key="plan_output",
)

# ─── Agent 3: Writer ─────────────────────────────────────────────────
writer_agent = LlmAgent(
    name="writer",
    model=MODEL_ID,
    instruction="""You are an expert **Technical Writer & Content Creator**. Your job is to 
produce the full written deliverable based on the plan.

**Plan to execute:**
{plan_output}

**Original Research:**
{research_output}

**Your Process:**
1. Review the plan and research carefully
2. Write the complete deliverable, following the plan's structure
3. Ensure every planned item is addressed
4. Use professional formatting and clear language

**Writing Standards:**
- **Professional tone**: Clear, confident, and well-organized
- **Rich formatting**: Use headers (##), bullet points, numbered lists, bold, and tables
- **Comprehensive**: Cover every point from the plan
- **Practical**: Include actionable details, examples, and specific recommendations
- **Well-structured**: Use logical flow with clear sections and transitions
- **Evidence-based**: Reference findings from the research where relevant

**Document Structure:**
1. Title & Introduction
2. Main content sections (following the plan)
3. Detailed recommendations
4. Implementation guidance
5. Conclusion & Next Steps

Write the COMPLETE document — no placeholders, no "TBD" sections. This should be 
a polished, ready-to-use deliverable.""",
    description="Writes the full deliverable document based on the plan",
    tools=[],
    output_key="written_output",
)

# ─── Agent 4: Reviewer ───────────────────────────────────────────────
reviewer_agent = LlmAgent(
    name="reviewer",
    model=MODEL_ID,
    instruction="""You are an expert **Quality Reviewer & Editor**. Your job is to review, 
refine, and polish the written deliverable to ensure it's exceptional.

**Document to review:**
{written_output}

**Original plan for completeness check:**
{plan_output}

**Original research for accuracy check:**
{research_output}

**Your Review Process:**

### 1. Completeness Check
- Does the document cover ALL items from the plan?
- Are there any gaps or missing sections?

### 2. Accuracy Check  
- Are facts and claims consistent with the research?
- Are recommendations realistic and well-supported?

### 3. Quality Check
- Is the writing clear, professional, and engaging?
- Is the formatting consistent and well-organized?
- Are there any grammatical or spelling errors?

### 4. Enhancement
- Add any missing details that would strengthen the document
- Improve weak sections with better examples or explanations
- Ensure smooth transitions between sections

**Your Output:**
First, provide a brief **Review Summary** (3-5 bullet points on what you improved).
Then, provide the **COMPLETE final polished document** — not just the changes, but the 
entire document with all improvements applied.

Use the `save_to_file` tool to save the final document with an appropriate filename.

The final document should be exceptional — something that would impress a senior executive 
or client.""",
    description="Reviews, refines, and delivers the final polished output",
    tools=[save_to_file],
    output_key="review_output",
)

# ─── Pipeline: Sequential Agent ──────────────────────────────────────
root_agent = SequentialAgent(
    name="task_planner_pipeline",
    description=(
        "A multi-agent task planning pipeline that researches, plans, writes, "
        "and reviews deliverables. Four specialized agents work in sequence — "
        "each builds on the previous agent's work."
    ),
    sub_agents=[
        researcher_agent,
        planner_agent,
        writer_agent,
        reviewer_agent,
    ],
)
